tags policy, and controlls button

+++ app/Http/Controllers/TagController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
@@ -107,6 +108,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         $oldName = $tag->name;
         $tag->delete();
         return $this->index()->with(
